+++ src/public/mathfield-element.ts
@@ -7,6 +7,7 @@ import {
     Offset,
     Range,
     Selection,
+    FindOptions,
 } from './mathfield';
 import { MathfieldErrorCode, ParseMode, ParserErrorCode, Style } from './core';
 
@@ -638,20 +639,29 @@ export class MathfieldElement extends HTMLElement implements Mathfield {
      * Updates the style (color, bold, italic, etc...) of the selection or sets
      * the style to be applied to future input.
      *
-     * If there is a selection, the style is applied to the selection
+     * If there is no selection and no range is specified, the style will
+     * apply to the next character typed.
      *
-     * If the selection already has this style, it is removed.
+     * If a range is specified, the style is applied to the range, otherwise,
+     * if there is a selection, the style is applied to the selection.
      *
-     * If the selection has the style partially applied (i.e. only some
-     * sections), it is removed from those sections, and applied to the
-     * entire selection.
+     * If the operation is 'toggle' and the range already has this style,
+     * remove it. If the range
+     * has the style partially applied (i.e. only some sections), remove it from
+     * those sections, and apply it to the entire range.
      *
-     * If there is no selection, the style will apply to the next character typed.
+     * If the operation is 'set', the style is applied to the range,
+     * whether it already has the style or not.
+     *
+     * The default operation is 'set'.
      *
      * @category Accessing and changing the content
      */
-    applyStyle(style: Style): void {
-        return this.#mathfield?.applyStyle(style);
+    applyStyle(
+        style: Style,
+        options?: Range | { range?: Range; operation?: 'set' | 'toggle' }
+    ): void {
+        return this.#mathfield?.applyStyle(style, options);
     }
 
     /**
@@ -685,8 +695,8 @@ export class MathfieldElement extends HTMLElement implements Mathfield {
      * An array is always returned, but it has no element if there are no
      * matching items.
      */
-    find(value: string | RegExp): Range[] {
-        return this.#mathfield?.find(value) ?? [];
+    find(value: string | RegExp, options?: FindOptions): Range[] {
+        return this.#mathfield?.find(value, options) ?? [];
     }
     /**
      * Custom elements lifecycle hooks